Rewrite to store log entries in a single table
This way we need much less magic (eg. no models are dynamically
created), and we can rely mostly on Django's contenttypes framework and
use generic foreign keys to point back to the audit logged object.

This also removes the need to manually add migration operations, as we
have a field type we can use to detect models that should have audit
logging added.